+++ NeuroRighter/Plotting/SnipGridGraph.cs
@@ -40,7 +40,7 @@ namespace NeuroRighter
         VertexDeclaration vDec;
         List<VertexPositionColor[]> lines; //Lines to be plotted
         List<VertexPositionColor[]> gridLines; //Grid lines
-        int[] idx; //Index to points in 'lines'
+        short[] idx; //Index to points in 'lines'
         private static readonly short[] gridIdx = { 0, 1 }; //Index to points in gridLines
 
         //Constants for text rendering
@@ -54,6 +54,7 @@ namespace NeuroRighter
         private double voltageRange; //in volts
         private double timeRange; //in seconds
         private Vector2 voltageTimeLabelCoords;
+   
 
         private int waveformsPerPlot;
 
@@ -62,7 +63,6 @@ namespace NeuroRighter
         {
             lock (lockObject)
             {
-
                 this.numRows = numRows; this.numCols = numColumns; this._isSpikeWaveformPlot = isSpikeWaveformPlot;
                 this.numSamplesPerPlot = numSamplesPerPlot;
                 this.waveformsPerPlot = waveformsPerPlot;
@@ -70,13 +70,14 @@ namespace NeuroRighter
                 lines = new List<VertexPositionColor[]>(numCols * numRows * waveformsPerPlot);
                 for (int i = 0; i < numCols * numRows * waveformsPerPlot; ++i)
                     lines.Add(new VertexPositionColor[numSamplesPerPlot]);
-                idx = new int[numSamplesPerPlot];
+                idx = new short[numSamplesPerPlot];
 
 
                 gridLines = new List<VertexPositionColor[]>(numRows + numCols - 2);
 
                 for (int i = 0; i < numRows + numCols - 2; ++i) gridLines.Add(new VertexPositionColor[2]);
-                for (int i = 0; i < idx.Length; ++i) idx[i] = i;
+                for (short i = 0; i < idx.Length; ++i) 
+                    idx[i] = i;
 
                 this.timeRange = timeRange;
                 this.voltageRange = voltageRange;
